// src/app/models/user.model.ts
export class User {
    departmentId: string;
    department: string;
    code: string;
    errorDesc: string;
    errorCode: string;
    userId: string;
    userType: string;
    applicableTo: string;
    fName: string;
    utName: string;
    userName: string;
    iUserId: string;
    userTag: string;
    imagePath: string;
}